Test cases addded
